refactor: DTO 디렉토리 생성 후 이름 변경
