+++ src/main/java/homeTry/exerciseList/dto/response/ExerciseResponse.java
@@ -1,20 +1,20 @@
-package homeTry.exerciseList.dto;
+package homeTry.exerciseList.dto.response;
 
 import homeTry.exerciseList.model.entity.Exercise;
 import homeTry.exerciseList.model.entity.ExerciseHistory;
 import homeTry.exerciseList.model.entity.ExerciseTime;
 import java.time.Duration;
 
-public record ExerciseDTO(
+public record ExerciseResponse(
     Long exerciseId,
     String exerciseName,
     Duration exerciseTime,
     boolean isActive
 ) {
 
-    public static ExerciseDTO of(Exercise exercise, Duration exerciseTime,
+    public static ExerciseResponse of(Exercise exercise, Duration exerciseTime,
         boolean isActive) {
-        return new ExerciseDTO(
+        return new ExerciseResponse(
             exercise.getExerciseId(),
             exercise.getExerciseName(),
             exerciseTime,
@@ -22,11 +22,11 @@ public record ExerciseDTO(
         );
     }
 
-    public static ExerciseDTO from(ExerciseHistory history) {
+    public static ExerciseResponse from(ExerciseHistory history) {
         return of(history.getExercise(), history.getExerciseHistoryTime(), false);
     }
 
-    public static ExerciseDTO from(ExerciseTime exerciseTime) {
+    public static ExerciseResponse from(ExerciseTime exerciseTime) {
         return of(exerciseTime.getExercise(), exerciseTime.getExerciseTime(),
             exerciseTime.isActive());
     }
